Feature/mq 1771

See merge request medeq/api!31

+++ Modules/Product/Database/Migrations/2021_04_27_122215_create_product_variants_table.php
@@ -17,8 +17,8 @@ class CreateProductVariantsTable extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained();
             $table->string('name');
-            $table->unsignedInteger('price')->nullable();
-            $table->unsignedInteger('previous_price')->nullable();
+            $table->unsignedBigInteger('price')->nullable();
+            $table->unsignedBigInteger('previous_price')->nullable();
             $table->foreignId('currency_id')->nullable()->constrained();
             $table->boolean('is_price_visible')->default(false);
             $table->boolean('is_enabled')->default(true);
